[FIX] dms: Change size field type from integer to float to prevent 2147483647 limit from postgresql.
TT33935

+++ dms/models/dms_file.py
@@ -107,7 +107,7 @@ class File(models.Model):
         compute="_compute_mimetype", string="Type", readonly=True, store=True
     )
 
-    size = fields.Integer(string="Size", readonly=True)
+    size = fields.Float(string="Size", readonly=True)
 
     checksum = fields.Char(string="Checksum/SHA1", readonly=True, index=True)
